requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
pypdf
python-dotenv
tiktoken
fastapi
uvicorn
pydantic
sentence-transformers
